requirements.txt
pyxel>=2.0.0
pyserial>=3.5
Pillow>=10.0.0
scapy>=2.5.0
netifaces>=0.11.0
bleak>=0.21.0
dbus-python>=1.3.2
LoRaRF>=1.4.0
cryptography>=41.0.0
PyNaCl>=1.5.0
